pyproject.toml
[project]
name = "podcast-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.20",
    "apscheduler==4.0.0a6",
    "feedparser>=6.0",
    "httpx>=0.27",
    "pydantic-ai>=0.0.49",
    "python-dotenv>=1.0",
    "python-telegram-bot>=20.0",
    "python-ulid>=2.0",
    "faster-whisper>=1.0",
    "groq>=0.8",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff>=0.15.6",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = ["ignore::telegram.warnings.PTBUserWarning"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"
